requirements.txt
Faker==40.8.0
joblib==1.5.2
jsonlines==4.0.0
numpy==2.4.2
PyYAML==6.0.3
scikit_learn==1.8.0
django